pyproject.toml
[project]
name = "green-wealth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "openai>=1.64.0",
    "pyotp>=2.9.0",
    "python-dotenv>=1.0.1",
    "sendgrid>=6.11.0",
    "wtforms>=3.2.1",
]

[dependency-groups]
dev = [
    "ruff>=0.9.7",
]
